Added a Master class
The master class is responsible for holding all objects we only need to
instantiate once. For example: AmazonS3Client and the Cognito
credentials should all be initialized here and just referenced from
here. Had to add an entry to the AndroidManifest.xml to include it.
Usage is as follows:

Master m = (Master)context.getApplicationContext();
this.s3Client = m.getS3Client();
...